+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- `IPython.display.Markdown` now renders as markdown instead of plain text.
+
 ## 0.4.0 - 2026-01-26
 
 ### Fixed